Rename dsp namespace
